Combine vector space with RAM/ROM object space
Vectors now use the same address space as regular objects. A benefit to
this is that vectors and objects can potentially be allocated from the
same memory pool. Another benefit is that the compiler can potentially
store read-only vectors in ROM.

One drawback to this is that vectors can no longer occupy a full 32KiB
of memory because the 32KiB address space is now shared with simple
objects, ROM objects, and RAM objects. On architectures with a total of
only 32KiB of usable memory (eg, Arduino), this is not an issue.