INCOMPLETE! reimplement ports and peek chars
A port is now a pair consisting of a file descriptor and peek char. This
data structure may expand in the future to hold more information about
the port, such as its type (input/output, binary/text) and whether it's
open or closed.

There also seems to be a bug in the compiler: if *stdout* is explicitly
referenced, (current-output-port) returns the current output port, which
is equal to *stdout* by default. Otherwise (current-output-port) returns #f.